Remove unused model structs and folder code, rename domain.rs
