--- crates/bitwarden/src/platform/folders/mod.rs
@@ -1,13 +1,0 @@
-mod create;
-mod delete;
-mod list;
-mod update;
-
-pub(crate) use create::create_folder;
-pub use create::FolderCreateRequest;
-pub(crate) use delete::delete_folder;
-pub use delete::FolderDeleteRequest;
-pub(crate) use list::list_folders;
-pub use list::{FolderView, FoldersResponse};
-pub(crate) use update::update_folder;
-pub use update::FolderUpdateRequest;
