+++ crates/bitwarden/src/state/mod.rs
@@ -1,3 +1,3 @@
-pub mod domain;
+pub mod models;
 pub mod state;
 pub(crate) mod state_service;
