+++ crates/bitwarden/src/state/state_service.rs
@@ -1,11 +1,11 @@
-use std::{collections::HashMap, marker::PhantomData};
+use std::marker::PhantomData;
 
 use serde::{de::DeserializeOwned, Serialize};
-use uuid::Uuid;
 
-use crate::error::Result;
-
-use super::{domain::*, state::State};
+use crate::{
+    error::Result,
+    state::{models::*, state::State},
+};
 
 #[derive(Clone, Copy)]
 pub(crate) struct ServiceDefinition<T: Serialize + DeserializeOwned> {
@@ -16,7 +16,7 @@ pub(crate) struct ServiceDefinition<T: Serialize + DeserializeOwned> {
 }
 
 impl<T: Serialize + DeserializeOwned> ServiceDefinition<T> {
-    const fn new(namespace: &'static str) -> Self {
+    pub(crate) const fn new(namespace: &'static str) -> Self {
         let _type = PhantomData;
         Self { namespace, _type }
     }
@@ -25,12 +25,7 @@ impl<T: Serialize + DeserializeOwned> ServiceDefinition<T> {
 pub(crate) const KEYS_SERVICE: ServiceDefinition<Option<Keys>> = ServiceDefinition::new("keys");
 pub(crate) const PROFILE_SERVICE: ServiceDefinition<Option<Profile>> =
     ServiceDefinition::new("profile");
-pub(crate) const CIPHERS_SERVICE: ServiceDefinition<HashMap<Uuid, Cipher>> =
-    ServiceDefinition::new("ciphers");
-pub(crate) const FOLDERS_SERVICE: ServiceDefinition<HashMap<Uuid, Folder>> =
-    ServiceDefinition::new("folders");
 pub(crate) const AUTH_SERVICE: ServiceDefinition<Auth> = ServiceDefinition::new("auth");
-pub(crate) const SETTINGS_SERVICE: ServiceDefinition<Settings> = ServiceDefinition::new("settings");
 
 pub(crate) struct StateService<'a, T: Serialize + DeserializeOwned + Default> {
     state: &'a State,
